requirements.txt
amqp==2.5.2
asgiref==3.2.3
asn1crypto==0.24.0
aspy.yaml==1.1.2
attrs==19.1.0
backports.csv==1.0.7
bcrypt==3.1.6
billiard==3.6.1.0
cairocffi==0.9.0
CairoSVG==2.2.1
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
cfgv==1.4.0
chardet==3.0.4
coverage==4.5.2
cryptography==2.7
cssselect==1.1.0
cssselect2==0.2.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-braces==1.13.0
django-cachalot==2.1.0
django-celery-beat==1.5.0
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-dotenv==1.4.2
django-extensions==2.1.9
django-extra-views==0.12.0
django-filter==2.1.0
django-filters==0.2.1
django-fontawesome==1.0
django-import-export==1.2.0
django-jsonforms==1.1.0
django-model-utils==3.1.2
django-multiupload==0.5.2
django-npm==1.0.0
django-redis==4.10.0
django-rest-framework==0.1.0
django-select2==7.0.3
django-stronghold==0.3.0
django-tables2==2.0.6
django-watson==1.5.2
djangorestframework==3.9.4
entrypoints==0.3
et-xmlfile==1.0.1
fabric==2.4.0
flake8==3.7.5
graphviz==0.13
gunicorn==19.9.0
html5lib==1.0.1
httplib2==0.14.0
identify==1.2.1
idna==2.8
importlib-metadata==1.3.0
importlib-resources==1.0.2
inflect==2.1.0
invoke==1.2.0
jdcal==1.4.1
jsonfield==2.0.2
jsonschema==3.0.1
kombu==4.6.7
lxml==4.4.1
mccabe==0.6.1
more-itertools==8.0.2
nodeenv==1.3.3
numpy==1.17.0
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
ordereddict==1.1
pandas==0.25.0
paramiko==2.4.2
Pillow==6.0.0
pre-commit==1.14.3
protobuf==3.8.0
psycopg2==2.7.7
psycopg2-binary==2.8.2
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.0
Pygments==2.4.2
PyNaCl==1.3.0
Pyphen==0.9.5
pyrsistent==0.15.2
python-crontab==2.4.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
reportlab==3.5.23
requests==2.22.0
requests-oauthlib==1.2.0
sentry-sdk==0.9.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
tinycss==0.4
tinycss2==0.6.1
toml==0.10.0
treelib==1.5.5
unicodecsv==0.14.1
urllib3==1.25.3
vine==1.3.0
virtualenv==16.3.0
WeasyPrint==44
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
